pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "geometric_kernels"
authors = [
    {name = "The GeometricKernels Contributors", email = "geometric-kernels@googlegroups.com"},
]
description="A Python Package offering geometric kernels in NumPy, TensorFlow, PyTorch, and Jax."
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords=[
    "geometric-kernels",
]
requires-python = ">=3.8"
dependencies = [
    "backends>=1.5.4",
    "einops",
    "geomstats",
    "numpy>=1.16",
    "opt-einsum",
    "plum-dispatch>=2.2.0",
    "potpourri3d",
    "robust_laplacian",
    "scipy>=1.3",
    "spherical-harmonics-basis",
    "sympy~=1.13",
]
version="0.2.2"

[project.urls]
Documentation = "https://geometric-kernels.github.io/"
Source = "https://github.com/geometric-kernels/GeometricKernels"

[tool.isort]
profile = "black"
skip_glob = [
    "geometric_kernels/__init__.py",
]
known_third_party = [
    "lab",
]

[tool.mypy]
ignore_missing_imports = true
strict_optional = false
allow_redefinition = true


[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
